+++ source/src/View/AddAdditionalMissionBehaviourView.cs
@@ -1,5 +1,6 @@
 ﻿using RTSCamera.Config;
 using System.Collections.Generic;
+using RTSCamera.Logic;
 using RTSCamera.Patch;
 using RTSCamera.Patch.CircularFormation;
 using TaleWorlds.MountAndBlade;