+++ source/src/View/FlyCameraMissionView.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using System.Reflection;
 using RTSCamera.Config;
+using RTSCamera.Logic;
 using TaleWorlds.Core;
 using TaleWorlds.Engine;
 using TaleWorlds.Engine.Screens;