+++ source/src/View/HideHUDView.cs
@@ -1,4 +1,5 @@
 ﻿using RTSCamera.Config;
+using RTSCamera.Logic;
 using TaleWorlds.Engine;
 using TaleWorlds.Engine.Screens;
 using TaleWorlds.InputSystem;